Refactor, add report processor
- Cleaned up Makefile
- Add Allure reporter
- Break spec/targets.js into multiple files
- Add test for annual targets
- Create multiple WDIO config files
