+++ test/js/package.json
@@ -3,6 +3,7 @@
 		"test": "mocha --require babel-register --es_staging --harmony"
 	},
 	"devDependencies": {
+		"allure-commandline": "^2.5.0",
 		"babel-cli": "^6.26.0",
 		"babel-preset-env": "^1.6.1",
 		"babel-preset-es2015": "^6.24.1",
@@ -11,6 +12,7 @@
 		"jsdoc": "^3.5.5",
 		"mocha": "^4.1.0",
 		"selenium-webdriver": "^3.6.0",
+		"wdio-allure-reporter": "^0.3.2",
 		"wdio-mocha-framework": "^0.5.12",
 		"wdio-selenium-standalone-service": "0.0.9",
 		"wdio-spec-reporter": "^0.1.3",
